Implemented BookContainer.Install()
Fallback is set on BookContainer initialisation.

For few of the more common interfaces available in EpubContainer
made APIs in BookContainer as well:

    BookContainer.HadErrors
    BookContainer.Fallback
    BookContainer.Items

Fixed: OpenPackageFile.ContainsItem now checks for the value
rather than the key, which is the correct behaviour as it
should be looking in the paths, not the ids.