BrowserEngine: Use resources instead of content
The required files for the browser engine used to be compiled as
content. This lead to external assemblies, e.g. AlbiteTests, to
need to have a second copy of those files, creating unnecessary
redundancy.

These files are now set as resources and are part of the AlbiteREADER
assembly, thus are stored only once.