Set up the SystemTray and the ApplicationBar
Now the SystemTray and the ApplicationBar are on top of the
application page (opacity < 1.0) and thus won't affect the layout.
This means that there won't be relayouts when made (in)visible.

Also, the BackgroundColor of the ApplicationBar is set to the
background color of the page as it looks better this way.