Fixed #95: Current position not retained
[Problem]
The problem is due to the fact that no nodes get approved in getTextLocation():

  if (content.length === 0) {
    // No nodes. Perhaps an empty document?
    // Default to the beginning of the document
    return makeLocation();
  }

getCurrentPage() and getPageCount() work fine though.

[Solution]
Wait for some frames before declaring "loaded".

This fixes the problem with range.getBoundingClientRect() returning
empty boxes for text nodes, while the page is still getting rendered.

Also, this remedies the issue of seeing a blank page before the user
gets to the right one (i.e. the user sees the agent scrolling).

The idea is that the first couple of frames will be slow, as the
user agent would still be laying out the content and trying to render.
However, the remainder of the frames will be fired at around 1/60 ms,
so adding around 100ms is not a huge problem. Making sure the (buggy)
user agent has finished layout/render is capital though.