Go to chapter + fragment doesn't work on some books
The problem was that the BoundClientRect's width could be zero,
which is perfectly valid if the anchor didn't contain any content,
which was checked for and skipped.

Now box's width is not taken into account.

However, sometimes it can still be somewhat inaccurate because
the anchor could be left on the page before the actual content,
e.g. the title, thus giving the experience that it's not going
to the right page. There is not much to do, though.