JSClient: DomLocation now is XML-correct
Exported DomLocation from JSClient didn't account for albite_root and
albite_start, thus it was not correct outside JSClient.

This was not a problem if it was used only internally.

However, with the addition of Narration, the location is used in other
modules as well.

The right place to correct this is in JSClient.

DomLocation now is now under version control - when imported into
JSClient it is transformed into internal DomLocation depending on its
version.