TextElement: Replace NBSPs to SPs as well
Sometimes, the "Back" button on NarrationPage appear not to work,
i.e. it gets one to the same text, i.e. the experience is as though
one has reached the first page of the first chapter.

This was due to the fact that before that text there was a
TextElement that was only white space characters, yet was not
filtered away because of the usage of NBSPs.

This lead to the mentioned experience. If the user was to
quickly tap "Back" several times, once they might be able to
tap it before the "reached" event of the next text block
was fired, thus "unblocking" the situate.

Silly, how something so different in nature, could produce
this particular problem.