App: Fixed the type of the dependency property for BookmarkControl
When BookmarkControl is a part of a List, it was not created for
items not using Bookmark, but other IBookmark.

Corrected the type of the Bookmark dependency property from
Bookmark to IBookmark.

The problem was difficult to identify as there were no
exceptions thrown, only no items would get added to the list